fix: move WriterDraftService to internalapi package for external access
- 将WriterDraftService从service/internal移动到service/internalapi
- 更新handlers.go和router.go中的import路径
- 解决Go internal包访问限制问题
- 确保API层可以正常访问Service层
